# api/openapi_chatgpt_operations_v016.yaml
openapi: 3.1.0
info:
  title: SPPG Vendor and Inventory Operations
  version: 0.16.0
  description: Parse vendor WhatsApp invoices, reconcile PO/receiving/invoice/reject layers, manage operational stock, and confirm vendor payment evidence. No finance ledger transaction is created automatically by these operations.
servers:
  - url: https://sppg-finance-gpt-site-production-5b7d.up.railway.app
security:
  - bearerAuth: []
paths:
...
  /v1/vendor-payables/from-receipt:
    post:
      operationId: processSppgVendorPayableFromReceipt
      summary: Preview or commit reconciled vendor payable
      description: Always use commit=false first. Keep PO qty, received qty, invoice qty, rejected qty, and payable qty separate. vendor_cost_price must be explicit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorPayableRequest'
      responses:
        '200':
          description: Payable preview or commit result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorPayableResponse'
        '409':
          description: Item or quantity reconciliation is ambiguous
  /v1/vendor-payables:
    get:
      operationId: searchSppgVendorPayables
      summary: Search vendor payables
      parameters:
        - in: query
          name: site
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: vendor
          schema: {type: string}
        - in: query
          name: status
          schema: {type: string, enum: [UNPAID, PARTIAL, PAID, CANCELLED]}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 500, default: 200}
      responses:
        '200':
          description: Matching payables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayableSearchResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    VendorCostLine:
      type: object
      required: [vendor_cost_price]
      properties:
        goods_receipt_item_id: {type: [integer, 'null']}
        item_name: {type: [string, 'null']}
        vendor_cost_price: {type: number, minimum: 0}
        invoiced_qty: {type: [number, 'null'], minimum: 0}
        rejected_qty: {type: number, minimum: 0, default: 0}
    VendorPayableRequest:
      type: object
      required: [site, purchase_order_id, goods_receipt_id, commit, lines]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        purchase_order_id: {type: integer}
        goods_receipt_id: {type: integer}
        invoice_number: {type: [string, 'null']}
        invoice_date: {type: [string, 'null'], format: date}
        due_date: {type: [string, 'null'], format: date}
        evidence_uri: {type: [string, 'null']}
        commit: {type: boolean, default: false}
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/VendorCostLine'
    VendorPayableLine:
      type: object
      properties:
        goods_receipt_item_id: {type: integer}
        purchase_order_item_id: {type: [integer, 'null']}
        item_name: {type: string}
        accepted_qty: {type: number}
        invoiced_qty: {type: number}
        rejected_qty: {type: number}
        payable_qty: {type: number}
        unit: {type: [string, 'null']}
        planned_qty: {type: [number, 'null']}
        po_qty: {type: [number, 'null']}
        planning_price: {type: [number, 'null']}
        po_price: {type: [number, 'null']}
        vendor_cost_price: {type: number}
        gross_line_total: {type: number}
        reject_amount: {type: number}
        line_total: {type: number}
        invoice_vs_po_variance: {type: [number, 'null']}
        invoice_vs_receipt_variance: {type: number}
        warnings:
          type: array
          items: {type: string}
    VendorPayableResponse:
      type: object
      properties:
        committed: {type: boolean}
        canCommit: {type: boolean}
        duplicate: {type: boolean}
        vendorInvoiceId: {type: [integer, 'null']}
        site: {type: string}
        purchaseOrderId: {type: integer}
        poCode: {type: string}
        goodsReceiptId: {type: integer}
        vendorCode: {type: string}
        payableStatus: {type: string}
        grossAmount: {type: number}
        rejectDeduction: {type: number}
        netAmount: {type: number}
        financeTransactionCreated: {type: boolean}
        warnings:
          type: array
          items: {type: string}
        lines:
          type: array
          items:
            $ref: '#/components/schemas/VendorPayableLine'
    PayableRecord:
      type: object
      properties:
        vendor_invoice_id: {type: integer}
        vendor_code: {type: string}
        site: {type: string}
        purchase_order_id: {type: [integer, 'null']}
        goods_receipt_id: {type: [integer, 'null']}
        invoice_number: {type: [string, 'null']}
        gross_amount: {type: number}
        reject_deduction: {type: number}
        net_amount: {type: number}
        payable_status: {type: string}
    PayableSearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PayableRecord'
...